fix: split oversized project and intent files (fixes #302)
