+++ internal/web/chat_intent.go
@@ -4,15 +4,10 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"net/http"
-	"os"
-	"os/exec"
 	"path/filepath"
-	"regexp"
-	"sort"
 	"strings"
 	"time"
 
@@ -460,7 +455,6 @@ func normalizeSystemActionForExecution(action *SystemAction, fallbackText string
 			action.Params["input_mode"] = chatInputModeText
 		}
 	}
-	_ = fallbackText
 	return action
 }
 
@@ -506,8 +500,7 @@ func safeFindToken(raw string) string {
 		}
 	}
 	clean := strings.TrimSpace(out.String())
-	clean = strings.Trim(clean, "/")
-	return clean
+	return strings.Trim(clean, "/")
 }
 
 func buildOpenCanvasFallbackPlan(text string) []*SystemAction {
@@ -557,15 +550,11 @@ func buildOpenCanvasFallbackPlan(text string) []*SystemAction {
 	return []*SystemAction{
 		{
 			Action: "shell",
-			Params: map[string]interface{}{
-				"command": command,
-			},
+			Params: map[string]interface{}{"command": command},
 		},
 		{
 			Action: "open_file_canvas",
-			Params: map[string]interface{}{
-				"path": systemActionLastShellPathPlaceholder,
-			},
+			Params: map[string]interface{}{"path": systemActionLastShellPathPlaceholder},
 		},
 	}
 }
@@ -595,9 +584,7 @@ func ensureOpenCanvasTerminalAction(actions []*SystemAction) []*SystemAction {
 	repaired = append(repaired, actions...)
 	repaired = append(repaired, &SystemAction{
 		Action: "open_file_canvas",
-		Params: map[string]interface{}{
-			"path": systemActionLastShellPathPlaceholder,
-		},
+		Params: map[string]interface{}{"path": systemActionLastShellPathPlaceholder},
 	})
 	return repaired
 }
@@ -687,7 +674,6 @@ func (a *App) classifyIntentPlanWithLLM(ctx context.Context, text string) ([]*Sy
 	if err != nil {
 		return nil, err
 	}
-
 	if requiresOpenCanvas && !planContainsAction(actions, "open_file_canvas") {
 		previousPlanJSON := "null"
 		if len(actions) > 0 {
@@ -712,7 +698,6 @@ func (a *App) classifyIntentPlanWithLLM(ctx context.Context, text string) ([]*Sy
 			return nil, nil
 		}
 	}
-
 	if len(actions) == 0 {
 		return nil, nil
 	}
@@ -756,9 +741,7 @@ func (a *App) classifyAndExecuteSystemAction(ctx context.Context, sessionID stri
 			return message, payloads, true
 		}
 		if isExplicitDangerDecline(trimmedText) {
-			return "Canceled dangerous action.", []map[string]interface{}{{
-				"type": "confirmation_canceled",
-			}}, true
+			return "Canceled dangerous action.", []map[string]interface{}{{"type": "confirmation_canceled"}}, true
 		}
 	}
 
@@ -834,8 +817,6 @@ func (a *App) classifyAndExecuteSystemAction(ctx context.Context, sessionID stri
 				}
 				return message, payloads, true
 			}
-			// Route tool actions through Qwen plan decoding so one prompt can trigger
-			// multiple coordinated actions (e.g., shell + open_file_canvas).
 			if normalized.Action != "shell" && normalized.Action != "open_file_canvas" {
 				if message, payloads, ok := tryExecutePlan([]*SystemAction{normalized}); ok {
 					return message, payloads, true
@@ -847,720 +828,4 @@ func (a *App) classifyAndExecuteSystemAction(ctx context.Context, sessionID stri
 		return "", nil, false
 	}
 	return "", nil, false
-}
-
-func firstShellPathFromOutput(output string) string {
-	lines := strings.Split(output, "\n")
-	for _, line := range lines {
-		candidate := strings.TrimSpace(line)
-		if candidate == "" {
-			continue
-		}
-		if candidate == "(no output)" {
-			continue
-		}
-		candidate = strings.TrimPrefix(candidate, "./")
-		if candidate == "." || candidate == ".." {
-			continue
-		}
-		return candidate
-	}
-	return ""
-}
-
-type shellPathCandidate struct {
-	Title         string
-	HintScore     int
-	HiddenPenalty int
-	NoisyPenalty  int
-	Depth         int
-	Length        int
-}
-
-var quotedTextPattern = regexp.MustCompile(`"([^"]+)"|'([^']+)'`)
-
-func normalizeOpenHintToken(raw string) string {
-	token := strings.ToLower(strings.TrimSpace(raw))
-	token = strings.Trim(token, " \t\r\n`'\".,:;!?()[]{}<>")
-	token = strings.TrimPrefix(token, "./")
-	token = strings.Trim(token, "/")
-	token = strings.ReplaceAll(token, "\\", "/")
-	return token
-}
-
-func extractOpenRequestHints(text string) []string {
-	trimmed := strings.TrimSpace(text)
-	if trimmed == "" {
-		return nil
-	}
-	stopwords := map[string]struct{}{
-		"open": {}, "show": {}, "display": {}, "read": {}, "view": {}, "edit": {},
-		"the": {}, "a": {}, "an": {}, "please": {}, "file": {}, "files": {},
-		"on": {}, "in": {}, "at": {}, "to": {}, "from": {}, "for": {},
-		"canvas": {}, "project": {}, "this": {}, "that": {}, "my": {},
-	}
-	addable := func(token string, seen map[string]struct{}, out *[]string) {
-		token = normalizeOpenHintToken(token)
-		if token == "" {
-			return
-		}
-		if _, blocked := stopwords[token]; blocked {
-			return
-		}
-		if len(token) < 3 && !strings.Contains(token, ".") && !strings.Contains(token, "/") {
-			return
-		}
-		if _, exists := seen[token]; exists {
-			return
-		}
-		seen[token] = struct{}{}
-		*out = append(*out, token)
-	}
-
-	hints := make([]string, 0, 8)
-	seen := map[string]struct{}{}
-
-	// Quoted file/path fragments are usually strong user hints.
-	for _, match := range quotedTextPattern.FindAllStringSubmatch(trimmed, -1) {
-		for _, group := range match[1:] {
-			if strings.TrimSpace(group) == "" {
-				continue
-			}
-			addable(group, seen, &hints)
-		}
-	}
-
-	fields := strings.Fields(strings.ToLower(trimmed))
-	verbs := map[string]struct{}{"open": {}, "show": {}, "display": {}, "read": {}, "view": {}, "edit": {}}
-	for i, field := range fields {
-		verb := normalizeOpenHintToken(field)
-		if _, isVerb := verbs[verb]; !isVerb {
-			continue
-		}
-		for j := i + 1; j < len(fields) && j <= i+6; j++ {
-			addable(fields[j], seen, &hints)
-		}
-	}
-
-	// Extract explicit file/path-like tokens anywhere in the request.
-	for _, field := range strings.Fields(trimmed) {
-		token := normalizeOpenHintToken(field)
-		if token == "" {
-			continue
-		}
-		if strings.Contains(token, ".") || strings.Contains(token, "/") {
-			addable(token, seen, &hints)
-			base := normalizeOpenHintToken(filepath.Base(token))
-			addable(base, seen, &hints)
-			stem := normalizeOpenHintToken(strings.TrimSuffix(base, filepath.Ext(base)))
-			addable(stem, seen, &hints)
-		}
-	}
-	return hints
-}
-
-func scoreShellPathCandidate(title string, hints []string) int {
-	if len(hints) == 0 {
-		return 0
-	}
-	cleanTitle := filepath.ToSlash(strings.ToLower(strings.TrimSpace(title)))
-	if cleanTitle == "" {
-		return 0
-	}
-	base := strings.ToLower(strings.TrimSpace(filepath.Base(cleanTitle)))
-	stem := strings.TrimSuffix(base, filepath.Ext(base))
-	score := 0
-	for _, rawHint := range hints {
-		hint := normalizeOpenHintToken(rawHint)
-		if hint == "" {
-			continue
-		}
-		hintBase := strings.ToLower(strings.TrimSpace(filepath.Base(hint)))
-		hintStem := strings.TrimSuffix(hintBase, filepath.Ext(hintBase))
-		switch {
-		case cleanTitle == hint || base == hintBase:
-			score += 120
-		case stem != "" && (stem == hintBase || stem == hintStem):
-			score += 90
-		case strings.HasSuffix(cleanTitle, "/"+hintBase):
-			score += 70
-		case strings.Contains(base, hintBase):
-			score += 55
-		case strings.Contains(cleanTitle, hint):
-			score += 35
-		}
-	}
-	return score
-}
-
-func shellPathNoisyPenalty(title string) int {
-	clean := filepath.ToSlash(strings.ToLower(strings.TrimSpace(title)))
-	if clean == "" {
-		return 0
-	}
-	penalty := 0
-	segments := strings.Split(clean, "/")
-	for _, segment := range segments {
-		switch strings.TrimSpace(segment) {
-		case "node_modules", ".venv", "vendor", "dist", "build", "target", "gcc-build", "__pycache__":
-			penalty += 2
-		}
-	}
-	return penalty
-}
-
-func selectBestShellPathFromOutput(cwd, output string, hints []string) string {
-	root := strings.TrimSpace(cwd)
-	if root == "" {
-		return firstShellPathFromOutput(output)
-	}
-	lines := strings.Split(output, "\n")
-	candidates := make([]shellPathCandidate, 0, len(lines))
-	for _, line := range lines {
-		candidate := strings.TrimSpace(line)
-		if candidate == "" || candidate == "(no output)" {
-			continue
-		}
-		candidate = strings.TrimPrefix(candidate, "./")
-		candidate = strings.TrimSpace(candidate)
-		if candidate == "" || candidate == "." || candidate == ".." {
-			continue
-		}
-		absPath, canvasTitle, err := resolveCanvasFilePath(root, candidate)
-		if err != nil {
-			continue
-		}
-		info, statErr := os.Stat(absPath)
-		if statErr != nil || info.IsDir() {
-			continue
-		}
-		title := filepath.ToSlash(strings.TrimSpace(canvasTitle))
-		if title == "" {
-			continue
-		}
-		segments := strings.Split(title, "/")
-		hiddenPenalty := 0
-		for _, segment := range segments {
-			seg := strings.TrimSpace(segment)
-			if strings.HasPrefix(seg, ".") {
-				hiddenPenalty++
-			}
-		}
-		depth := strings.Count(title, "/")
-		candidates = append(candidates, shellPathCandidate{
-			Title:         title,
-			HintScore:     scoreShellPathCandidate(title, hints),
-			HiddenPenalty: hiddenPenalty,
-			NoisyPenalty:  shellPathNoisyPenalty(title),
-			Depth:         depth,
-			Length:        len(title),
-		})
-	}
-	if len(candidates) == 0 {
-		return firstShellPathFromOutput(output)
-	}
-	sort.SliceStable(candidates, func(i, j int) bool {
-		left := candidates[i]
-		right := candidates[j]
-		if left.HintScore != right.HintScore {
-			return left.HintScore > right.HintScore
-		}
-		if left.HiddenPenalty != right.HiddenPenalty {
-			return left.HiddenPenalty < right.HiddenPenalty
-		}
-		if left.NoisyPenalty != right.NoisyPenalty {
-			return left.NoisyPenalty < right.NoisyPenalty
-		}
-		if left.Depth != right.Depth {
-			return left.Depth < right.Depth
-		}
-		return left.Length < right.Length
-	})
-	return candidates[0].Title
-}
-
-func resolveRootTopLevelFile(root, rel string) (string, bool) {
-	if strings.TrimSpace(root) == "" || strings.TrimSpace(rel) == "" {
-		return "", false
-	}
-	cleanRel := strings.TrimSpace(filepath.Base(filepath.Clean(rel)))
-	if cleanRel == "" || cleanRel == "." || cleanRel == ".." {
-		return "", false
-	}
-	abs := filepath.Clean(filepath.Join(root, cleanRel))
-	if info, err := os.Stat(abs); err == nil && !info.IsDir() {
-		return filepath.ToSlash(cleanRel), true
-	}
-	entries, err := os.ReadDir(root)
-	if err != nil {
-		return "", false
-	}
-	for _, entry := range entries {
-		if entry.IsDir() {
-			continue
-		}
-		name := strings.TrimSpace(entry.Name())
-		if name == "" {
-			continue
-		}
-		if strings.EqualFold(name, cleanRel) {
-			return filepath.ToSlash(name), true
-		}
-	}
-	return "", false
-}
-
-func preferTopLevelSiblingPath(cwd, candidate string) string {
-	cleanCandidate := filepath.ToSlash(strings.TrimSpace(candidate))
-	if cleanCandidate == "" {
-		return cleanCandidate
-	}
-	base := strings.TrimSpace(filepath.Base(cleanCandidate))
-	if base == "" || base == "." || base == ".." {
-		return cleanCandidate
-	}
-	root := strings.TrimSpace(cwd)
-	if root == "" {
-		return cleanCandidate
-	}
-
-	if resolved, ok := resolveRootTopLevelFile(root, base); ok {
-		return resolved
-	}
-
-	ext := strings.TrimSpace(filepath.Ext(base))
-	if ext == "" {
-		stem := strings.TrimSpace(base)
-		variants := []string{
-			stem + ".md",
-			stem + ".markdown",
-			stem + ".txt",
-			stem + ".rst",
-			stem + ".adoc",
-		}
-		for _, variant := range variants {
-			if resolved, ok := resolveRootTopLevelFile(root, variant); ok {
-				return resolved
-			}
-		}
-	}
-	return cleanCandidate
-}
-
-func (a *App) executeSystemActionPlan(sessionID string, session store.ChatSession, userText string, actions []*SystemAction) (string, []map[string]interface{}, error) {
-	actions = enforceRoutingPolicy(userText, actions)
-	if len(actions) == 0 {
-		return "", nil, errors.New("action plan is empty")
-	}
-	if guardMessage, guardPayloads, blocked := a.guardDangerousSystemActionPlan(sessionID, userText, actions); blocked {
-		return guardMessage, guardPayloads, nil
-	}
-	return a.executeSystemActionPlanUnsafe(sessionID, session, userText, actions)
-}
-
-func (a *App) executeSystemActionPlanUnsafe(sessionID string, session store.ChatSession, userText string, actions []*SystemAction) (string, []map[string]interface{}, error) {
-	if len(actions) == 0 {
-		return "", nil, errors.New("action plan is empty")
-	}
-	messages := make([]string, 0, len(actions))
-	payloads := make([]map[string]interface{}, 0, len(actions))
-	lastShellPath := ""
-	targetProject, targetErr := a.systemActionTargetProject(session)
-	targetCWD := ""
-	if targetErr == nil {
-		targetCWD = strings.TrimSpace(targetProject.RootPath)
-		if targetCWD == "" {
-			targetCWD = strings.TrimSpace(a.cwdForProjectKey(targetProject.ProjectKey))
-		}
-	}
-	requestHints := extractOpenRequestHints(userText)
-	for _, action := range actions {
-		if action == nil {
-			continue
-		}
-		resolved := &SystemAction{
-			Action: action.Action,
-			Params: map[string]interface{}{},
-		}
-		for key, value := range action.Params {
-			resolved.Params[key] = value
-		}
-		if resolved.Action == "open_file_canvas" {
-			path := systemActionOpenPath(resolved.Params)
-			if strings.EqualFold(strings.TrimSpace(path), systemActionLastShellPathPlaceholder) {
-				if strings.TrimSpace(lastShellPath) == "" {
-					return "", nil, errors.New("open_file_canvas requires a resolved shell path")
-				}
-				resolved.Params["path"] = preferTopLevelSiblingPath(targetCWD, lastShellPath)
-			}
-		}
-		message, payload, err := a.executeSystemAction(sessionID, session, resolved)
-		if err != nil {
-			return "", nil, err
-		}
-		if strings.TrimSpace(message) != "" {
-			messages = append(messages, strings.TrimSpace(message))
-		}
-		if payload != nil {
-			payloads = append(payloads, payload)
-			payloadType := strings.TrimSpace(fmt.Sprint(payload["type"]))
-			payloadOutput := strings.TrimSpace(fmt.Sprint(payload["output"]))
-			if strings.EqualFold(payloadType, "shell") && payloadOutput != "" && payloadOutput != "<nil>" {
-				lastShellPath = selectBestShellPathFromOutput(targetCWD, payloadOutput, requestHints)
-			}
-		}
-	}
-	if len(messages) == 0 {
-		messages = append(messages, "Done.")
-	}
-	return strings.Join(messages, "\n\n"), payloads, nil
-}
-
-func (a *App) systemActionTargetProject(session store.ChatSession) (store.Project, error) {
-	projectKey := strings.TrimSpace(session.ProjectKey)
-	if projectKey != "" {
-		project, err := a.store.GetProjectByProjectKey(projectKey)
-		if err == nil && !isHubProject(project) {
-			return project, nil
-		}
-	}
-	return a.hubPrimaryProject()
-}
-
-func truncateSystemActionOutput(text string, maxBytes int) string {
-	if maxBytes <= 0 {
-		maxBytes = systemActionShellOutputLimit
-	}
-	if len(text) <= maxBytes {
-		return text
-	}
-	if maxBytes <= 24 {
-		return text[:maxBytes]
-	}
-	return text[:maxBytes] + "\n...(truncated)"
-}
-
-type shellCommandExecution struct {
-	Output   string
-	ExitCode int
-	TimedOut bool
-	RunErr   error
-}
-
-func executeShellCommand(command string, cwd string) shellCommandExecution {
-	commandCtx, cancel := context.WithTimeout(context.Background(), systemActionShellTimeout)
-	defer cancel()
-
-	cmd := exec.CommandContext(commandCtx, "bash", "-lc", command)
-	cmd.Dir = cwd
-
-	var output bytes.Buffer
-	cmd.Stdout = &output
-	cmd.Stderr = &output
-	runErr := cmd.Run()
-
-	rawOutput := strings.TrimSpace(output.String())
-	rawOutput = truncateSystemActionOutput(rawOutput, systemActionShellOutputLimit)
-	if rawOutput == "" {
-		rawOutput = "(no output)"
-	}
-
-	if errors.Is(commandCtx.Err(), context.DeadlineExceeded) {
-		return shellCommandExecution{
-			Output:   rawOutput,
-			ExitCode: -1,
-			TimedOut: true,
-			RunErr:   runErr,
-		}
-	}
-
-	exitCode := 0
-	if runErr != nil {
-		var exitErr *exec.ExitError
-		if errors.As(runErr, &exitErr) {
-			exitCode = exitErr.ExitCode()
-		}
-	}
-
-	return shellCommandExecution{
-		Output:   rawOutput,
-		ExitCode: exitCode,
-		TimedOut: false,
-		RunErr:   runErr,
-	}
-}
-
-func suggestShellCommandRetry(command string, output string) (string, string, bool) {
-	cleanCommand := strings.TrimSpace(command)
-	if cleanCommand == "" || !strings.Contains(cleanCommand, "jq") {
-		return "", "", false
-	}
-	cleanOutput := strings.TrimSpace(output)
-	if cleanOutput == "" {
-		return "", "", false
-	}
-	lowerOutput := strings.ToLower(cleanOutput)
-	if !strings.Contains(lowerOutput, "jq: error: syntax error") || !strings.Contains(lowerOutput, "compile error") {
-		return "", "", false
-	}
-	lines := strings.Split(cleanOutput, "\n")
-	for _, line := range lines {
-		candidate := strings.TrimSpace(line)
-		if candidate == "" || !strings.HasPrefix(candidate, ".") {
-			continue
-		}
-		if !(strings.HasSuffix(candidate, "}") || strings.HasSuffix(candidate, "]")) {
-			continue
-		}
-		fixedCandidate := strings.TrimSuffix(strings.TrimSuffix(candidate, "}"), "]")
-		if strings.TrimSpace(fixedCandidate) == "" || fixedCandidate == candidate {
-			continue
-		}
-		fixedCommand := strings.Replace(cleanCommand, candidate, fixedCandidate, 1)
-		if fixedCommand == cleanCommand {
-			continue
-		}
-		return fixedCommand, fmt.Sprintf("fixed jq filter typo (%s -> %s)", candidate, fixedCandidate), true
-	}
-	return "", "", false
-}
-
-func (a *App) executeSystemAction(sessionID string, session store.ChatSession, action *SystemAction) (string, map[string]interface{}, error) {
-	if action == nil {
-		return "", nil, errors.New("system action is required")
-	}
-	switch action.Action {
-	case "switch_project":
-		targetName := systemActionStringParam(action.Params, "name")
-		project, err := a.hubFindProjectByName(targetName)
-		if err != nil {
-			return "", nil, err
-		}
-		activated, err := a.activateProject(project.ID)
-		if err != nil {
-			return "", nil, err
-		}
-		return fmt.Sprintf("Switched to %s.", activated.Name), map[string]interface{}{
-			"type":       "switch_project",
-			"project_id": activated.ID,
-		}, nil
-	case "switch_workspace":
-		workspace, err := a.resolveWorkspaceReference(session.ProjectKey, systemActionWorkspaceRef(action.Params))
-		if err != nil {
-			return "", nil, err
-		}
-		if err := a.store.SetActiveWorkspace(workspace.ID); err != nil {
-			return "", nil, err
-		}
-		return fmt.Sprintf("Switched to workspace %s.", workspace.Name), map[string]interface{}{
-			"type":         "switch_workspace",
-			"workspace_id": workspace.ID,
-			"name":         workspace.Name,
-			"dir_path":     workspace.DirPath,
-		}, nil
-	case "list_workspace_items":
-		workspace, err := a.resolveWorkspaceReference(session.ProjectKey, systemActionWorkspaceRef(action.Params))
-		if err != nil {
-			return "", nil, err
-		}
-		items, err := a.listOpenWorkspaceItems(workspace.ID)
-		if err != nil {
-			return "", nil, err
-		}
-		itemIDs := make([]int64, 0, len(items))
-		titles := make([]string, 0, len(items))
-		for _, item := range items {
-			itemIDs = append(itemIDs, item.ID)
-			titles = append(titles, item.Title)
-		}
-		return summarizeWorkspaceItems(workspace, items), map[string]interface{}{
-			"type":         "list_workspace_items",
-			"workspace_id": workspace.ID,
-			"name":         workspace.Name,
-			"dir_path":     workspace.DirPath,
-			"item_ids":     itemIDs,
-			"item_titles":  titles,
-			"item_count":   len(items),
-		}, nil
-	case "create_workspace_from_git":
-		workspace, repoURL, err := a.createWorkspaceFromGit(systemActionGitRepoURL(action.Params), systemActionGitTargetPath(action.Params))
-		if err != nil {
-			return "", nil, err
-		}
-		return fmt.Sprintf("Created workspace %s at %s.", workspace.Name, workspace.DirPath), map[string]interface{}{
-			"type":         "create_workspace_from_git",
-			"workspace_id": workspace.ID,
-			"name":         workspace.Name,
-			"dir_path":     workspace.DirPath,
-			"repo_url":     repoURL,
-		}, nil
-	case "switch_model":
-		targetProject, err := a.systemActionTargetProject(session)
-		if err != nil {
-			return "", nil, err
-		}
-		updated, err := a.updateProjectChatModel(
-			targetProject.ID,
-			systemActionStringParam(action.Params, "alias"),
-			systemActionStringParam(action.Params, "effort"),
-		)
-		if err != nil {
-			return "", nil, err
-		}
-		return fmt.Sprintf(
-				"Model for %s set to %s (%s).",
-				updated.Name,
-				updated.ChatModel,
-				updated.ChatModelReasoningEffort,
-			), map[string]interface{}{
-				"type":       "switch_model",
-				"project_id": updated.ID,
-				"alias":      updated.ChatModel,
-				"effort":     updated.ChatModelReasoningEffort,
-			}, nil
-	case "toggle_silent":
-		return "Toggled silent mode.", map[string]interface{}{"type": "toggle_silent"}, nil
-	case "toggle_live_dialogue":
-		return "Toggled Live Dialogue.", map[string]interface{}{"type": "toggle_live_dialogue"}, nil
-	case "cancel_work":
-		activeCanceled, queuedCanceled := a.cancelChatWork(sessionID)
-		total := activeCanceled + queuedCanceled
-		return fmt.Sprintf("Canceled %d running task(s).", total), nil, nil
-	case "show_status":
-		status, err := a.fetchCodexStatusMessage(session.ProjectKey)
-		if err != nil {
-			return "", nil, err
-		}
-		return status, nil, nil
-	case "capture_idea":
-		return a.captureIdeaItem(session, action)
-	case "make_item", "delegate_item", "snooze_item", "split_items":
-		return a.createConversationItem(sessionID, session, action)
-	case "create_github_issue", "create_github_issue_split":
-		return a.createGitHubIssueFromConversation(sessionID, session, action)
-	case "shell":
-		targetProject, err := a.systemActionTargetProject(session)
-		if err != nil {
-			return "", nil, err
-		}
-		command := systemActionShellCommand(action.Params)
-		if command == "" {
-			return "", nil, errors.New("shell command is required")
-		}
-		cwd := strings.TrimSpace(targetProject.RootPath)
-		if cwd == "" {
-			cwd = strings.TrimSpace(a.cwdForProjectKey(targetProject.ProjectKey))
-		}
-		if cwd == "" {
-			return "", nil, errors.New("shell cwd is not available")
-		}
-		execResult := executeShellCommand(command, cwd)
-		if execResult.TimedOut {
-			return fmt.Sprintf("Shell command timed out after %s.\n\n%s", systemActionShellTimeout, execResult.Output), map[string]interface{}{
-				"type":       "shell",
-				"command":    command,
-				"cwd":        cwd,
-				"exit_code":  -1,
-				"timed_out":  true,
-				"output":     execResult.Output,
-				"project_id": targetProject.ID,
-			}, nil
-		}
-
-		if execResult.RunErr != nil && execResult.ExitCode == 0 {
-			return "", nil, execResult.RunErr
-		}
-
-		if execResult.ExitCode != 0 {
-			if fixedCommand, fixReason, retry := suggestShellCommandRetry(command, execResult.Output); retry {
-				retryResult := executeShellCommand(fixedCommand, cwd)
-				if !retryResult.TimedOut && retryResult.RunErr == nil && retryResult.ExitCode == 0 {
-					return fmt.Sprintf("Shell command auto-corrected (%s).\n\n%s", fixReason, retryResult.Output), map[string]interface{}{
-						"type":                "shell",
-						"command":             fixedCommand,
-						"original_command":    command,
-						"cwd":                 cwd,
-						"exit_code":           0,
-						"output":              retryResult.Output,
-						"project_id":          targetProject.ID,
-						"auto_corrected":      true,
-						"auto_correct_reason": fixReason,
-					}, nil
-				}
-			}
-			return fmt.Sprintf("Shell command failed (exit %d).\n\n%s", execResult.ExitCode, execResult.Output), map[string]interface{}{
-				"type":       "shell",
-				"command":    command,
-				"cwd":        cwd,
-				"exit_code":  execResult.ExitCode,
-				"output":     execResult.Output,
-				"project_id": targetProject.ID,
-			}, nil
-		}
-		return execResult.Output, map[string]interface{}{
-			"type":       "shell",
-			"command":    command,
-			"cwd":        cwd,
-			"exit_code":  execResult.ExitCode,
-			"output":     execResult.Output,
-			"project_id": targetProject.ID,
-		}, nil
-	case "open_file_canvas":
-		targetProject, err := a.systemActionTargetProject(session)
-		if err != nil {
-			return "", nil, err
-		}
-		rawPath := systemActionOpenPath(action.Params)
-		if rawPath == "" {
-			return "", nil, errors.New("open_file_canvas path is required")
-		}
-		cwd := strings.TrimSpace(targetProject.RootPath)
-		if cwd == "" {
-			cwd = strings.TrimSpace(a.cwdForProjectKey(targetProject.ProjectKey))
-		}
-		if cwd == "" {
-			return "", nil, errors.New("open_file_canvas cwd is not available")
-		}
-		absPath, canvasTitle, err := resolveCanvasFilePath(cwd, rawPath)
-		if err != nil {
-			return "", nil, err
-		}
-		info, err := os.Stat(absPath)
-		if err != nil {
-			return "", nil, err
-		}
-		if info.IsDir() {
-			return "", nil, fmt.Errorf("path %q is a directory", rawPath)
-		}
-		if info.Size() > systemActionOpenFileSizeLimit {
-			return "", nil, fmt.Errorf("file %q exceeds %d bytes", rawPath, systemActionOpenFileSizeLimit)
-		}
-		contentBytes, err := os.ReadFile(absPath)
-		if err != nil {
-			return "", nil, err
-		}
-		canvasSessionID := strings.TrimSpace(a.canvasSessionIDForProject(targetProject))
-		if canvasSessionID == "" {
-			return "", nil, errors.New("canvas session is not available")
-		}
-		port, ok := a.tunnels.getPort(canvasSessionID)
-		if !ok {
-			return "", nil, fmt.Errorf("no active MCP tunnel for project %q", targetProject.Name)
-		}
-		if _, err := a.mcpToolsCall(port, "canvas_artifact_show", map[string]interface{}{
-			"session_id":       canvasSessionID,
-			"kind":             "text",
-			"title":            canvasTitle,
-			"markdown_or_text": string(contentBytes),
-		}); err != nil {
-			return "", nil, err
-		}
-		return fmt.Sprintf("Opened %s on canvas.", canvasTitle), map[string]interface{}{
-			"type":       "open_file_canvas",
-			"path":       canvasTitle,
-			"project_id": targetProject.ID,
-		}, nil
-	default:
-		return "", nil, fmt.Errorf("unsupported action: %s", action.Action)
-	}
 }
